Improve readability and refactoring
Current code has magic numbers and duplicate methods that
accomplish the same task of checking for input
It was also not at the same level of abstraction

Consolidate Duplicate Conditional Fragments by removing
the switch statement in the main method. This also removed the
magic numbers that were in the main method.
Refactor code by extracting method as the println statements
from the marking and unmarking of tasks can be grouped together.

This is done according to the SE best practices of improving
readability and refactoring