Added "Mark as Done".
Added Task class to store the description of added messages and its marked status.
Added ability to return the string of the task represented with its marked status in brackets in front.
Added commands to mark and unmark with the 2nd argument in the input delimited by spaces to be the 1-based index of the task to be marked.
Replaced messages with tasks.