Add Level-4: ToDo, Event, Deadline
ToDo is a sub class of Task that represents activities that do not
require a deadline.
Deadline is a sub class of Task that represents an activity that needs
to be completed by a specified deadline.
Event is a sub class of Deadline that represents and activity that
starts at a specific date/time and ends at a specific date/time.