Allow Duke to handle exceptions
Duke is operating in an ideal world where there are no invalid inputs.

Users might not input commands in the correct format and Duke
needs to be able to handle it these exceptions

Add Duke Exceptions class to handle the various exceptions that can
occur when a user uses Duke. Duke will print an error message when
user inputs a task type that it does not recognise.