Added "Level 4. ToDos, Events, Deadlines"
Added 3 new commands of todo(), deadline(by), and event(from, to) that the user can type into the console.
Added 3 subclasses of Task to allow for the 3 new commands.
Added prefixes 'T', 'D' and 'E' in the returned display of the task to correspond to the type of task being referred to.
Added printAddedTask function to allow easier display of tasks in list
Changed name of project to DukeNUS.
Changed the welcome message logo
Removed the default command to add task when no command is specified.