Reformat constructions of Task related classes.
Currently when creating `Deadline`, `Event` and `Todo classes`, functions that
are used to process user input commands like `createDeadline()`,
`createEvent()` are called in the `Duck.class`.

After modification, these parts are moved to the constructors of the
respective classes.

The implementation of creating these three classes from input arguments
are also modified to use index of `by` or `from`, `to`, making the code
logic easier to read.

The name of the personal assisstant is changed to Tohru.

Text-UI-Test is updated.