Added "A-Jar" Create a JAR file.
Added functionality to run the program from a jar file.
Added function to store the tasks in tasks ArrayList in data/dukeNUS.txt, that will be retrieved by future instances of the program.
Added another parameter to all Task classes to allow functionality with the TasksDataWrite class and for the saved file to carry information on the done status of the task.