Added "A-JavaDoc" JavaDoc.
Added JavaDoc comments to almost all functions.
Changed display of date of tasks to be in yyyy-MM-dd to be more consistent with the inputs and the saving.
Changed the delimiters when writing to the file to include more than just a '/'.
Changed the way a saved file is read with the new delimiters.